# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TTR ONE — ERP Platform API
  version: 0.1.0
  description: |
    REST API for **TTR ONE**, a multi-tenant SaaS ERP for Uzbekistan
    (Node/Fastify + PostgreSQL/Prisma backend, Nuxt 3 SPA frontend).

    ## Conventions
    - **Base URL**: all endpoints live under `/api/v1`.
    - **Auth**: JWT Bearer. Obtain an `accessToken` from `POST /auth/login`
      (15-minute TTL) and rotate it via `POST /auth/refresh`.
    - **Money**: all monetary amounts are integers in **minor units (tiyin)**,
      e.g. `1 000 000` = 10 000.00 UZS. Fields end in `Minor`.
    - **Quantities**: decimal strings, bounded to 6 decimal places.
    - **Multi-tenancy**: every record is scoped to the caller's tenant
      (derived from the token); there is no tenant id in the path.
    - **Errors**: uniform envelope `{ "error": { "code": "...", "message": "..." } }`.
    - **Pagination**: list endpoints accept `page`, `pageSize` (max 200),
      `sort`, `order` (`asc`/`desc`), `search`, and return a `meta` object
      `{ page, pageSize, total, totalPages }`.
  contact:
    name: TTR ONE
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - apiKeyAuth: []
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Log in with email + password (optional tenant slug and MFA code)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /studio/integrations/{provider}:
    patch:
      tags: [studio]
      summary: Connect / configure an integration
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config:
                  type: object
                  additionalProperties: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '404': { $ref: '#/components/responses/NotFound' }
  /studio/forms:
    get:
      tags: [studio]
      summary: List no-code custom forms
      responses: { '200': { description: Forms } }
    post:
      tags: [studio]
      summary: Create a custom form (field schema)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, name, fields]
              properties:
                key: { type: string }
                name: { type: string }
                description: { type: string }
                icon: { type: string }
                fields:
                  type: array
                  items: { $ref: '#/components/schemas/FormField' }
      responses: { '201': { description: Created } }
  /studio/forms/{id}/records:
    post:
      tags: [studio]
      summary: Submit a record against a form
      parameters: [{ $ref: '#/components/parameters/id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
                  additionalProperties: true
      responses: { '201': { description: Record created } }

  # ==================== PLATFORM ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access token from POST /auth/login, sent as `Authorization: Bearer <token>`.'
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'Developer API key (ttr_live_…) created in the app under Управление → API-ключи. Recommended for server-to-server integrations. May also be sent as `Authorization: Bearer ttr_live_…`.'
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties: { ok: { type: boolean, example: true } }
...
    Unauthorized:
      description: Missing or invalid credentials/token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: { type: object, nullable: true }
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        tenant: { type: string, description: 'Slug, when the email exists in multiple tenants' }
        mfaCode: { type: string }
...
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tenant:
          type: object
          properties:
            id: { type: string }
            slug: { type: string }
            name: { type: string }
        user: { $ref: '#/components/schemas/User' }
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        fullName: { type: string }
        tenantId: { type: string }
        roles:
          type: array
          items: { type: string }
        permissions:
          type: array
          items: { type: string }
        platformAdmin: { type: boolean }
...
    FormField:
      type: object
      required: [key, label]
      properties:
        key: { type: string }
        label: { type: string }
        type: { type: string, enum: [text, textarea, number, date, select, checkbox], default: text }
        required: { type: boolean, default: false }
        options:
          type: array
          items: { type: string }